models/catboost_core/requirements.txt
catboost==1.2.5
numpy>=1.26
pandas>=2.2
duckdb>=0.10
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
scikit-learn>=1.5
python-dateutil>=2.9
python-dotenv>=1.0

